requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==21.4.0
avalanche-lib==0.1.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==5.0.1
blinker==1.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
click==8.0.3
clingo==5.6.2
cloudpickle==2.2.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.4.2
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gdown==4.2.1
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
graphviz==0.19.1
grpcio==1.43.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.10.1
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.3.2
lime==0.2.0.1
LTNtorch==1.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
mycolorpy==1.5.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
problog==2.2.2
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.33
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
pydantic==1.10.4
pyDeprecate==0.3.1
Pygments==2.13.0
pyparsing==3.0.7
pyrsistent==0.19.2
PySDD==0.2.10
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.5.9
pytorchcv==0.0.67
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
quadprog==0.1.11
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.4
setproctitle==1.3.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.1
stack-data==0.6.1
submitit==1.4.5
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.17.0
threadpoolctl==3.1.0
tifffile==2022.2.9
tinycss2==1.2.1
torch==1.13.0
torch-tb-profiler==0.3.1
torchmetrics==0.7.0
torchvision==0.14.0
torchviz==0.0.2
tornado==6.2
tqdm==4.62.3
traitlets==5.5.0
typing_extensions==4.1.1
urllib3==1.26.8
wandb==0.13.11
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.0.2
widgetsnbextension==4.0.3
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0